Polish `CCStoreBenchmark`
- Make it easier to run a specific set of benchmarks
